Former-commit-id: 629cd4de9716b3ee37946c3add1dbde7444a620b

+++ framework/sofa/core/DataEngine.cpp
@@ -44,7 +44,7 @@ DataEngine::~DataEngine()
 void DataEngine::updateAllInputsIfDirty()
 {
     const DDGLinkContainer& inputs = DDGNode::getInputs();
-    for( unsigned i=0, iend=inputs.size() ; i<iend ; ++i )
+    for(size_t i=0, iend=inputs.size() ; i<iend ; ++i )
     {
 //        std::cerr<<"DataEngine::updateAllInputsIfDirty() updating "<<static_cast<core::objectmodel::BaseData*>(inputs[i])->getName()<<std::endl;
         static_cast<core::objectmodel::BaseData*>(inputs[i])->updateIfDirty();
